Improved the level decoding estimation.
This new algorithm is simpler and seems to be more resilient. It's based
on the observation that the samples values appear to increase
monotonically (or nearly so) up to the air/water boundary and then
decrease somewhat more rapidly from there.

The estimated level may still drift due to environmental conditions,
sometimes dropping overnight as it gets cooler and there doesn't seem
to be much in the data that could help us calibrate. It should be good
enough for at a glance estimates of how full the tank is but not for
precise flow metering or leak detection.